requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
fonttools==4.57.0
future==1.0.0
geocoder==1.38.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
ratelim==0.1.6
requests==2.32.3
seaborn==0.13.2
six==1.17.0
tkintermapview==1.29
tzdata==2025.2
urllib3==2.3.0